Create deduck stock endpoint and service
* create some exceptions for handling usecases of book catalog
* change Long to Integer type of stockQuantity

+++ src/main/java/com/sansarch/bookstore_catalog_service/dto/CreateBookInputDto.java
@@ -13,5 +13,5 @@ public class CreateBookInputDto {
     private String title;
     private String author;
     private BigDecimal price;
-    private Long stockAvailability;
+    private Integer stockAvailability;
 }
