+++ src/main/java/com/sansarch/bookstore_catalog_service/dto/EditBookInputDto.java
@@ -13,5 +13,5 @@ public class EditBookInputDto {
     private String title;
     private String author;
     private BigDecimal price;
-    private Long stockAvailability;
+    private Integer stockAvailability;
 }
