+++ src/main/java/com/sansarch/bookstore_catalog_service/entity/Book.java
@@ -28,5 +28,5 @@ public class Book {
     private BigDecimal price;
 
     @Column(nullable = false)
-    private Long stockAvailability;
+    private Integer stockAvailability;
 }
